ref: move modal to separate component
