+++ src/pages/users/ui/user-table/user-row/ui/BanModal.module.scss
